requirements.txt
streamlit>=1.31.0
pillow>=10.0.0
rembg>=2.0.0
opencv-python-headless>=4.9.0
